simplify drawer code
