+++ packages/slice-machine/pages/changes.tsx
@@ -18,8 +18,10 @@ import { isLoading } from "@src/modules/loading";
 import { LoadingKeysEnum } from "@src/modules/loading/types";
 import useSliceMachineActions from "@src/modules/useSliceMachineActions";
 import { SyncError } from "@src/models/SyncError";
-import DeleteDocumentsDrawer from "@components/DeleteDocumentsDrawer";
-import DeleteDocumentsDrawerOverLimit from "@components/DeleteDocumentsDrawer/DeleteDocumentsDrawerOverLimit";
+import {
+  SoftDeleteDocumentsDrawer,
+  HardDeleteDocumentsDrawer,
+} from "@components/DeleteDocumentsDrawer";
 
 const Changes: React.FunctionComponent = () => {
   const {
@@ -118,8 +120,8 @@ const Changes: React.FunctionComponent = () => {
         />
         {PageContent}
       </Box>
-      <DeleteDocumentsDrawer />
-      <DeleteDocumentsDrawerOverLimit />
+      <SoftDeleteDocumentsDrawer />
+      <HardDeleteDocumentsDrawer />
     </Container>
   );
 };
